+++ gazebo/sensors/ImuSensor.hh
@@ -14,12 +14,14 @@
  * limitations under the License.
  *
 */
-
-#ifndef _IMUSENSOR_HH_
-#define _IMUSENSOR_HH_
+#ifndef _GAZEBO_IMUSENSOR_HH_
+#define _GAZEBO_IMUSENSOR_HH_
 
 #include <vector>
 #include <string>
+#include <ignition/math/Pose3.hh>
+#include <ignition/math/Quaternion.hh>
+#include <ignition/math/Vector3.hh>
 
 #include "gazebo/physics/PhysicsTypes.hh"
 #include "gazebo/sensors/Sensor.hh"
@@ -63,16 +65,36 @@ namespace gazebo
 
       /// \brief Returns the angular velocity.
       /// \return Angular velocity.
-      public: math::Vector3 GetAngularVelocity() const;
+      /// \deprecated See AngularVelocity() function that returns an
+      /// ignition::math::Vector3d object.
+      public: math::Vector3 GetAngularVelocity() const GAZEBO_DEPRECATED(6.0);
 
       /// \brief Returns the imu linear acceleration
       /// \return Linear acceleration.
-      public: math::Vector3 GetLinearAcceleration() const;
+      /// \deprecated See LinearVelocity() function that returns an
+      /// ignition::math::Vector3d object.
+      public: math::Vector3 GetLinearAcceleration() const
+              GAZEBO_DEPRECATED(6.0);
 
       /// \brief get orientation of the IMU relative to the reference pose
       /// \return returns the orientation quaternion of the IMU relative to
       /// the imu reference pose.
-      public: math::Quaternion GetOrientation() const;
+      /// \deprecated See Orientation() function that returns an
+      /// ignition::math::Quaterniond object.
+      public: math::Quaternion GetOrientation() const GAZEBO_DEPRECATED(6.0);
+
+      /// \brief Returns the angular velocity.
+      /// \return Angular velocity.
+      public: ignition::math::Vector3d AngularVelocity() const;
+
+      /// \brief Returns the imu linear acceleration
+      /// \return Linear acceleration.
+      public: ignition::math::Vector3d LinearAcceleration() const;
+
+      /// \brief get orientation of the IMU relative to the reference pose
+      /// \return returns the orientation quaternion of the IMU relative to
+      /// the imu reference pose.
+      public: ignition::math::Quaterniond Orientation() const;
 
       /// \brief Sets the current pose as the IMU reference pose
       public: void SetReferencePose();
@@ -85,16 +107,16 @@ namespace gazebo
       private: void OnLinkData(ConstLinkDataPtr &_msg);
 
       /// \brief Imu reference pose
-      private: math::Pose referencePose;
+      private: ignition::math::Pose3d referencePose;
 
       /// \brief Save previous imu linear velocity for computing acceleration.
-      private: math::Vector3 lastLinearVel;
+      private: ignition::math::Vector3d lastLinearVel;
 
       /// \brief Imu linear acceleration
-      private: math::Vector3 linearAcc;
+      private: ignition::math::Vector3d linearAcc;
 
       /// \brief store gravity vector to be added to the imu output.
-      private: math::Vector3 gravity;
+      private: ignition::math::Vector3d gravity;
 
       /// \brief Imu data publisher
       private: transport::PublisherPtr pub;