+++ Changelog.md
@@ -3,6 +3,9 @@
 1. Added magnetometer sensor. A contribution from Andrew Symington.
     * [Pull request #1788](https://bitbucket.org/osrf/gazebo/pull-request/1788)
 
+1. Added altimeter sensor. A contribution from Andrew Symington.
+    * [Pull request #1792](https://bitbucket.org/osrf/gazebo/pull-request/1792)
+
 1. Implement more control options for log playback:
   1. Rewind: The simulation starts from the beginning.
   1. Forward: The simulation jumps to the end of the log file.